+++ osctiny/__init__.py
@@ -3,4 +3,4 @@ from .osc import Osc
 
 
 __all__ = ['Osc']
-__version__ = "0.1.7"
+__version__ = "0.1.8"